Fixed create test window directory, sidebar link name and popup component in Modal.
