+++ frontend/src/app/testWindow/page.tsx
@@ -1,10 +1,11 @@
 import { getServerAuthSession } from "@/utils/auth";
-import SchedulesPage from "./pageClient";
+import TestWindowPage from "./pageClient";
 import { AuthProvider } from "@/components/authProvider"
 import {Session} from "next-auth";
 import InstructorCoursesClient from "@/app/instructorCourses/pageClient";
 import CreateTestWindow from "@/app/_components/testWindow/CreateTestWindow";
 import TestWindowExample from "@/app/_components/testWindow/TestWindowExample";
+import TestWindowCalendar from "@/app/_components/UI/TestWindowCalendar";
 
 /**
  * Default session for users no logged in
@@ -23,13 +24,14 @@ const DEFAULT_SESSION: Session = {
  * injection into Client Side Page
  * @constructor
  */
-export default async function Schedule() {
+export default async function TestWindow() {
     // Session variable
     const session =  await getServerAuthSession() ?? DEFAULT_SESSION;
 
     return (
         <AuthProvider session={session}>
-            <SchedulesPage/>
+            <TestWindowPage/>
+            {/* <TestWindowCalendar/> */}
         </AuthProvider>
     );
 }
